requirements.txt
argon2-cffi==20.1.0
Django==3.1.3
django-allauth==0.43.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.1